refactor Watson2014, create utils.geometry
