+++ pulse2percept/models/__init__.py
@@ -11,8 +11,6 @@
 
 """
 from .base import NotBuiltError, Model, SpatialModel, TemporalModel
-from .watson2014 import (Watson2014ConversionMixin, dva2ret, ret2dva,
-                         Watson2014DisplacementMixin)
 from .scoreboard import ScoreboardModel, ScoreboardSpatial
 from .axon_map import AxonMapModel, AxonMapSpatial
 from .nanduri2012 import (Nanduri2012Model, Nanduri2012Spatial,
@@ -22,16 +20,12 @@ __all__ = [
     'AxonMapModel',
     'AxonMapSpatial',
     'Model',
-    'dva2ret',
     'Nanduri2012Model',
-    'Nanduri2012SpatialMixin',
-    'Nanduri2012TemporalMixin',
+    'Nanduri2012Spatial',
+    'Nanduri2012Temporal',
     'NotBuiltError',
-    'ret2dva',
     'ScoreboardModel',
     'ScoreboardSpatial',
     'SpatialModel',
     'TemporalModel',
-    'Watson2014ConversionMixin',
-    'Watson2014DisplacementMixin'
 ]
