+++ pulse2percept/models/_nanduri2012.pyx
@@ -7,6 +7,7 @@ from libc.math cimport(pow as c_pow, exp as c_exp, fabs as c_abs,
 
 ctypedef cnp.float32_t float32
 ctypedef cnp.int32_t int32
+ctypedef cnp.uint32_t uint32
 
 @cdivision(True)
 cdef inline float32 expit(float32 x) nogil:
@@ -40,16 +41,17 @@ cpdef spatial_fast(const float32[:, ::1] stim,
     xgrid, ygrid : 1D float32 array
         An array of x or y coordinates at which to calculate the spatial
         response (microns)
-    rho : float32
-        The rho parameter of the scoreboard model (microns): exponential decay
-        constant for the current spread
+    atten_a : float32
+        Nominator of the attentuation function (Eq.2 in the paper)
+    atten_n : float32
+        Exponent of the attenuation function's denominator (Eq.2 in the paper)
     thresh_percept : float32
         Spatial responses smaller than ``thresh_percept`` will be set to zero
 
     """
     cdef:
-        size_t idx_el, idx_time, idx_space, n_el, n_time, n_space
-        size_t idx_bright, n_bright
+        uint32 idx_el, idx_time, idx_space, n_el, n_time, n_space
+        uint32 idx_bright, n_bright
         float32[:, ::1] bright
         float32 px_bright, denom, d2c, d2e, amp
 
@@ -66,20 +68,28 @@ cpdef spatial_fast(const float32[:, ::1] stim,
         idx_space = idx_bright % n_space
         idx_time = idx_bright / n_space
 
+        # At each pixel to be rendered, we need to sum up the contribution of
+        # each electrode:
         px_bright = 0.0
         for idx_el in range(n_el):
             amp = stim[idx_el, idx_time]
             if c_abs(amp) > 0:
+                # Calculate current spread for this electrode, given by the
+                # distance to the electrode center (`d2c`) and an exponential
+                # attenuation:
                 d2c = (c_pow(xgrid[idx_space] - xel[idx_el], 2) +
                        c_pow(ygrid[idx_space] - yel[idx_el], 2))
                 if d2c < c_pow(rel[idx_el], 2):
+                    # On the electrode surface:
                     denom = atten_a + c_pow(zel[idx_el], atten_n)
-                    px_bright = px_bright + amp * atten_a / denom
                 else:
+                    # Away from the electrode surface, calculate the distance
+                    # to the electrode egde:
                     d2e = (c_pow(c_sqrt(d2c) - rel[idx_el], 2) +
                            c_pow(zel[idx_el], 2))
                     denom = atten_a + c_pow(c_sqrt(d2e), atten_n)
-                    px_bright = px_bright + amp * atten_a / denom
+                # Add the contribution of this electrode:
+                px_bright = px_bright + amp * atten_a / denom
         if c_abs(px_bright) < thresh_percept:
             px_bright = 0.0
         bright[idx_time, idx_space] = px_bright
@@ -89,7 +99,7 @@ cpdef spatial_fast(const float32[:, ::1] stim,
 @cdivision(True)
 cpdef temporal_fast(const float32[:, ::1] stim,
                     const float32[::1] t_stim,
-                    const size_t[::1] idx_t_percept,
+                    const uint32[::1] idx_t_percept,
                     float32 dt,
                     float32 tau1,
                     float32 tau2,
@@ -97,16 +107,46 @@ cpdef temporal_fast(const float32[:, ::1] stim,
                     float32 asymptote,
                     float32 shift,
                     float32 slope,
-                    float32 eps):
-    """Cython implementation of the Nanduri 2012 temporal model"""
+                    float32 eps,
+                    float32 thresh_percept):
+    """Cython implementation of the Nanduri 2012 temporal model
+
+    Parameters
+    ----------
+    stim : 2D float32 array
+        A ``Stimulus.data`` container that contains time points as rows and
+        spatial locations as columns. This is the output of the spatial model.
+        The time points are specified in ``t_stim``.
+    t_stim : 1D float32 array
+        The time points for ``stim`` above.
+    dt : float32
+        Sampling time step (seconds)
+    tau1: float32
+        Time decay constant for the fast leaky integrater (seconds).
+    tau2: float32
+        Time decay constant for the charge accumulation (seconds).
+    tau3: float32
+        Time decay constant for the slow leaky integrator (seconds).
+    eps: float32
+        Scaling factor applied to charge accumulation.
+    asymptote: float32
+        Asymptote of the logistic function used in the stationary nonlinearity
+        stage.
+    slope: float32
+        Slope of the logistic function in the stationary nonlinearity stage.
+    shift: float32
+        Shift of the logistic function in the stationary nonlinearity stage.
+    thresh_percept: float32
+        Below threshold, the percept has brightness zero.
+
+    """
     cdef:
         float32 ca, r1, r2, r3, max_r3, r4a, r4b, r4c
         float32 t_sim, amp, scale
         float32[::1] all_r3
         float32[:, ::1] percept
-        size_t idx_space, idx_sim, idx_stim, idx_frame
-        size_t n_space, n_stim, n_percept
-        int32 n_sim
+        uint32 idx_space, idx_sim, idx_stim, idx_frame
+        uint32 n_space, n_stim, n_percept, n_sim
 
     n_percept = len(idx_t_percept)  # Py overhead
     n_stim = len(t_stim)  # Py overhead
@@ -146,24 +186,27 @@ cpdef temporal_fast(const float32[:, ::1] stim,
             r3 = float_max(r1 - eps * r2, 0)
             # Store `r3` for Step 2:
             all_r3[idx_sim] = r3
-            # Find the largest `r3` across time:
+            # Find the largest `r3` across time for Step 2:
             if r3 > max_r3:
                 max_r3 = r3
 
-        # Step 2: Use `max_R3` to calculate the slow response:
+        # Step 2: Use `max_R3` from Step 1 to calculate the slow response for
+        # time points:
         r4a = 0.0
         r4b = 0.0
         r4c = 0.0
         idx_stim = 0
         idx_frame = 0
+        # Scaling factor depends on `max_r3` from Step 1:
         scale = asymptote * expit((max_r3 - shift) / slope) / max_r3
+        # We have to restart the loop over all simulation time steps from 0:
         for idx_sim in range(n_sim):
             t_sim = idx_sim * dt
             # Access the right stimulus frame (same as above):
             if idx_stim + 1 < n_stim:
                 if t_sim >= t_stim[idx_stim + 1]:
                     idx_stim = idx_stim + 1
-            # Slow response:
+            # Slow response (3-stage leaky integrator):
             r4a = r4a + dt * (all_r3[idx_sim] * scale - r4a) / tau3
             r4b = r4b + dt * (r4a - r4b) / tau3
             r4c = r4c + dt * (r4b - r4c) / tau3
@@ -172,6 +215,8 @@ cpdef temporal_fast(const float32[:, ::1] stim,
                 # output a percept. We compare `idx_sim` to `idx_t_percept`
                 # rather than `t_sim` to `t_percept` because there is no good
                 # (fast) way to compare two floating point numbers:
+                if c_abs(r4c) < thresh_percept:
+                    r4c = 0.0
                 percept[idx_frame, idx_space] = r4c
                 idx_frame = idx_frame + 1
