+++ pulse2percept/models/nanduri2012.py
@@ -1,11 +1,43 @@
-"""`Nanduri2012Model`"""
+"""`Nanduri2012Model`, `Nanduri2012Spatial`, `Nanduri2012Temporal`"""
 import numpy as np
 from .base import Model, SpatialModel, TemporalModel
 from ._nanduri2012 import spatial_fast, temporal_fast
+from ..utils import Curcio1990Transform
 
 
 class Nanduri2012Spatial(SpatialModel):
-    """Spatial model"""
+    """Spatial response model of [Nanduri2012]_
+
+    Implements the spatial response model described in [Nanduri2012]_, which
+    assumes that the spatial activation of retinal tissue is equivalent to the
+    "current spread" :math:`I`, described as a function of distance :math:`r`
+    from the center of the stimulating electrode:
+
+    .. math::
+
+        I(r) =
+        \\begin{cases}
+            \\frac{\\verb!atten_a!}{\\verb!atten_a! + (r-a)^\\verb!atten_n!}
+                & r > a \\\\
+            1 & r \\leq a
+        \\end{cases}
+
+    where :math:`a` is the radius of the electrode (see Eq.2 in the paper).
+
+    .. note::
+
+        Use this class if you just want the spatial response model.
+        Use :py:class:`~pulse2percept.models.Nanduri2012Model` if you want both
+        the spatial and temporal model.
+
+    Parameters
+    ----------
+    atten_a : float, optional, default: 14000.0
+        Nominator of the attentuation function
+    atten_n : float32, optional, default: 1.69
+        Exponent of the attenuation function's denominator
+
+    """
 
     def get_default_params(self):
         """Returns all settable parameters of the Nanduri model"""
@@ -16,18 +48,18 @@ class Nanduri2012Spatial(SpatialModel):
     def dva2ret(self, xdva):
         """Convert degrees of visual angle (dva) to retinal eccentricity (um)
 
-        Assumes that one degree of visual angle is equal to 288 um on the
+        Assumes that one degree of visual angle is equal to 280 um on the
         retina.
         """
-        return 288.0 * xdva
+        return Curcio1990Transform.dva2ret(xdva)
 
     def ret2dva(self, xret):
         """Convert retinal eccentricity (um) to degrees of visual angle (dva)
 
-        Assumes that one degree of visual angle is equal to 288 um on the
+        Assumes that one degree of visual angle is equal to 280 um on the
         retina.
         """
-        return xret / 288.0
+        return Curcio1990Transform.ret2dva(xret)
 
     def predict_spatial(self, implant, t):
         """Predicts the brightness at spatial locations"""
@@ -66,7 +98,41 @@ class Nanduri2012Spatial(SpatialModel):
 
 
 class Nanduri2012Temporal(TemporalModel):
-    """Temporal model"""
+    """Temporal model of [Nanduri2012]_
+
+    Implements the temporal response model described in [Nanduri2012]_, which
+    assumes that the temporal activation of retinal tissue is the output of a
+    linear-nonlinear model cascade (see Fig.6 in the paper).
+
+    .. note::
+
+        Use this class if you just want the temporal response model.
+        Use :py:class:`~pulse2percept.models.Nanduri2012Model` if you want both
+        the spatial and temporal model.
+
+    Parameters
+    ----------
+    dt : float, optional, default: 5 microseconds
+        Sampling time step (seconds)
+    tau1: float, optional, default: 0.42 ms
+        Time decay constant for the fast leaky integrater.
+    tau2: float, optional, default: 45.25 ms
+        Time decay constant for the charge accumulation.
+    tau3: float, optional, default: 26.25 ms
+        Time decay constant for the slow leaky integrator.
+    eps: float, optional, default: 8.73
+        Scaling factor applied to charge accumulation.
+    asymptote: float, optional, default: 14.0
+        Asymptote of the logistic function used in the stationary nonlinearity
+        stage.
+    slope: float, optional, default: 3.0
+        Slope of the logistic function in the stationary nonlinearity stage.
+    shift: float, optional, default: 16.0
+        Shift of the logistic function in the stationary nonlinearity stage.
+    thresh_percept: float, optional, default: 0
+        Below threshold, the percept has brightness zero.
+
+    """
 
     def get_default_params(self):
         base_params = super().get_default_params()
@@ -105,7 +171,7 @@ class Nanduri2012Temporal(TemporalModel):
 
         # Beware of floating point errors! 29.999 will be rounded down to 29
         # by np.uint, so we need to np.round it first:
-        idx_percept = np.uint64(np.round(t_percept / self.dt))
+        idx_percept = np.uint32(np.round(t_percept / self.dt))
         t_percept = idx_percept * self.dt
         if np.unique(idx_percept).size < t_percept.size:
             raise ValueError("All times 't' must be distinct multiples of "
@@ -114,17 +180,33 @@ class Nanduri2012Temporal(TemporalModel):
                              t_spatial.astype(np.float32),
                              idx_percept,
                              self.dt, self.tau1, self.tau2, self.tau3,
-                             self.asymptote, self.shift, self.slope, self.eps)
+                             self.asymptote, self.shift, self.slope, self.eps,
+                             self.thresh_percept)
 
 
 class Nanduri2012Model(Model):
-    """Nanduri et al. (2012) Model
+    """[Nanduri2012]_ Model
 
     Implements the model described in [Nanduri2012]_, where percepts are
     circular and their brightness evolves over time.
 
+    The model combines two parts:
+
+    *  :py:class:`~pulse2percept.models.Nanduri2012Spatial` is used to
+       calculate the spatial activation function, which is assumed to be
+       equivalent to the "current spread" described as a function of distance
+       from the center of the stimulating electrode (see Eq.2 in the paper).
+
+    *  :py:class:`~pulse2percept.models.Nanduri2012Temporal` is used to
+       calculate the temporal activation function, which is assumed to be the
+       output of a linear-nonlinear cascade model (see Fig.6 in the paper).
+
     Parameters
     ----------
+    atten_a : float, optional, default: 14000.0
+        Nominator of the attentuation function (Eq.2 in the paper)
+    atten_n : float32, optional, default: 1.69
+        Exponent of the attenuation function's denominator (Eq.2 in the paper)
     dt : float, optional, default: 5 microseconds
         Sampling time step (seconds)
     tau1: float, optional, default: 0.42 ms
@@ -133,7 +215,6 @@ class Nanduri2012Model(Model):
         Time decay constant for the charge accumulation.
     tau3: float, optional, default: 26.25 ms
         Time decay constant for the slow leaky integrator.
-        Default: 26.25 / 1000 s.
     eps: float, optional, default: 8.73
         Scaling factor applied to charge accumulation.
     asymptote: float, optional, default: 14.0
